requirements.txt
crewai>=1.0.0
crewai-tools>=0.1.6
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.7.1
streamlit>=1.32.0
